# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SofTeam Projects API",
        "description": "This API was made to handle Softeam project information",
        "version": "1.0.0"
    },
    "servers":[
        {
            "url": "http://localhost:3000",
            "description": "test server"
        }
    ],
    "paths": {
        "/collaborator": {
            "get": {
                "summary": "Return all collaborators working at Softeam",
                "description": "Return all collaborators working at Softeam",
                "tags": ["Collaborators"],
                "security": [{"bearerAuth": []}],
                "responses": {
                    "500": {
                        "description": "Error trying to get collaborators",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "examples": {
                                    "error": {
                                        "value": {
                                            "error": "Error trying to connect with database"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref":"#/components/schemas/collaborators"
                                    }
                                },
                                "examples": {
                                    "collaborators": {
                                        "value": [
                                            {
                                                "name": "collaborator 1",
                                                "email": "collaborator1@email.com",
                                                "role": "Back-end developer"
                                            },
                                            {
                                                "name": "collaborator 2",
                                                "email": "collaborator2@email.com",
                                                "role": "Front-end developer"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/collaborator/create": {
            "post": {
                "summary": "create one collaborator",
                "description": "create one collaborator",
                "tags": ["Collaborators"],
                "security": [{"bearerAuth": []}],
                "requestBody": {
                    "content":{
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/collaborators"
                            },
                            "examples": {
                                "collaborator": {
                                    "value": {
                                        "name": "collaborator",
                                        "email": "collaborator@email.com",
                                        "password": "password",
                                        "role": "Back-end developer"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "examples": {
                                    "error": {
                                        "value": {
                                            "error": "Insuficient Data!"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Collaborator not found!"
                    },
                    "500": {
                        "description": "Error trying to create collaborator",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "examples": {
                                    "error": {
                                        "value": {
                                            "error": "Error trying to connect with database"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Collaborator Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/collaborators"
                                },
                                "examples": {
                                    "collaborator": {
                                        "value": {
                                            "_id":"collaboratorID",
                                            "name": "collaborator",
                                            "email": "collaborator@email.com",
                                            "role": "Back-end developer"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/collaborator/update/{_id}": {
            "put": {
                "summary": "Update one existing collaborator",
                "description": "Update one existing collaborator",
                "tags": ["Collaborators"],
                "security": [{"bearerAuth": []}],
                "parameters": [
                    {
                        "name": "_id",
                        "in": "path",
                        "description": "ID to find collaborator",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content":{
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/collaborators"
                            },
                            "examples": {
                                "collaborator": {
                                    "value": {
                                        "name": "collaborator",
                                        "email": "collaborator@email.com",
                                        "role": "Back-end developer"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Insuficient Data!"
                    },
                    "404": {
                        "description": "Collaborator not found!"
                    },
                    "500": {
                        "description": "Error trying to update collaborator",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "examples": {
                                    "error": {
                                        "value": {
                                            "error": "Error trying to connect with database"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/collaborators"
                                },
                                "examples": {
                                    "collaborator": {
                                        "value": {
                                            "_id": "collaboratorID",
                                            "name": "collaborator",
                                            "email": "collaborator@email.com",
                                            "role": "Back-end developer"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/collaborator/delete": {
            "delete":{
                "summary": "Delete an existing collaborator",
                "description": "Delete an existing collaborator",
                "tags": ["Collaborators"],
                "security": [{"bearerAuth": []}],
                "parameters": [
                    {
                        "name": "_id",
                        "in": "query",
                        "description": "ID from the collaborator that you want to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Insuficient Data!"
                    },
                    "500": {
                        "description": "Error trying to delete collaborator",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "examples": {
                                    "error": {
                                        "value": {
                                            "error": "Error trying to connect with database"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "object",
                                        "$ref": "#/components/schemas/collaborators"
                                    },
                                    "examples": {
                                        "collaborator": {
                                            "value": {
                                                "_id": "collaboratorID",
                                                "name": "collaborator",
                                                "email": "collaborator@email.com",
                                                "role": "Back-end developer"
                                            }
                                        }
                                    }
                                }
                            }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            },
            "collaborators": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes" : {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}